change size_t to uint32_t in bam_header_t
This may cause issues on 64-bit big-endian machines. Reported and fixed by Paolo Emilio Mazzon.

+++ bam.h
@@ -89,7 +89,7 @@ typedef struct {
 	char **target_name;
 	uint32_t *target_len;
 	void *dict, *hash, *rg2lib;
-	size_t l_text, n_text;
+	uint32_t l_text, n_text;
 	char *text;
 } bam_header_t;
